Updated to display correctly on Linux - tested with Ubuntu 12.10. Created separate ToolConfig.txt files for each version. Added readme.txt to explain the differences between version1 and version2 branches.
git-svn-id: svn://svn.code.sf.net/p/xlights/code/trunk@115 ad4834c4-2333-46e5-bf01-adba983f9b38

+++ version2/xLights/NetInfo.cpp
@@ -41,7 +41,7 @@ int NetInfoClass::CalcAbsChannel(int NetNum, int NetCh)
 int NetInfoClass::GetTotChannels()
 {
     int tot=0;
-    for(int NetNum=0; NetNum < NetMaxChannel.size(); NetNum++) {
+    for(size_t NetNum=0; NetNum < NetMaxChannel.size(); NetNum++) {
         tot+=NetMaxChannel[NetNum];
     }
     return tot;
@@ -62,7 +62,7 @@ void NetInfoClass::GetAllChannelNames(wxArrayString& names)
 wxString NetInfoClass::GetChannelName(int ChNum)
 {
     int NetCh=ChNum;
-    for (int NetNum=0; NetNum < NetMaxChannel.size(); NetNum++) {
+    for (size_t NetNum=0; NetNum < NetMaxChannel.size(); NetNum++) {
         if (NetCh < NetMaxChannel[NetNum]) {
             return wxString::Format(wxT("Ch %d: Net %d #%d"),int(ChNum+1),int(NetNum+1),int(NetCh+1));
         }
